separate survey module and routes
