+++ app/collect/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint('survey', __name__, template_folder='templates')
 
-from app.collect import survey
+from app.collect import survey, routes
